requirements.txt
# LLM & Agentic Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# arXiv Scraping
arxiv>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# PDF Parsing
PyPDF2>=3.0.0

# Environment Variables
python-dotenv>=1.0.0

# Token Counting
tiktoken>=0.7.0

# UI
streamlit>=1.38.0
